Moving entity system subclasses into their own dir.
