+++ src/core/ecs/system/systems/animated_sprite_rendering_entity_system.h
@@ -1,10 +1,10 @@
 #ifndef ANIMATED_SPRITE_RENDERING_ENTITY_SYSTEM_H
 #define ANIMATED_SPRITE_RENDERING_ENTITY_SYSTEM_H
 
-#include "entity_system.h"
-#include "../component/components/transform2D_component.h"
-#include "../component/components/animated_sprite_component.h"
-#include "../../math/space_handler.h"
+#include "../entity_system.h"
+#include "../../component/components/transform2D_component.h"
+#include "../../component/components/animated_sprite_component.h"
+#include "../../../math/space_handler.h"
 
 class AnimatedSpriteRenderingEntitySystem : public EntitySystem {
   private:
