+++ src/core/ecs/system/systems/directional_light_rendering_entity_system.h
@@ -1,10 +1,10 @@
 #ifndef DIRECTIONAL_LIGHT_RENDERING_ENTITY_SYSTEM_H
 #define DIRECTIONAL_LIGHT_RENDERING_ENTITY_SYSTEM_H
 
-#include "entity_system.h"
-#include "../component/components/light3D_component.h"
-#include "../component/components/material_component.h"
-#include "../../math/space_handler.h"
+#include "../entity_system.h"
+#include "../../component/components/light3D_component.h"
+#include "../../component/components/material_component.h"
+#include "../../../math/space_handler.h"
 
 class DirectionalLightRenderingEntitySystem : public EntitySystem {
   private:
