+++ src/core/ecs/system/systems/text_rendering_entity_system.h
@@ -1,11 +1,11 @@
 #ifndef TEXT_RENDERING_ENTITY_SYSTEM_H
 #define TEXT_RENDERING_ENTITY_SYSTEM_H
 
-#include "entity_system.h"
-#include "../component/components/transform2D_component.h"
-#include "../component/components/text_label_component.h"
-#include "../../math/space_handler.h"
-#include "../../utils/string_util.h"
+#include "../entity_system.h"
+#include "../../component/components/transform2D_component.h"
+#include "../../component/components/text_label_component.h"
+#include "../../../math/space_handler.h"
+#include "../../../utils/string_util.h"
 
 class TextRenderingEntitySystem : public EntitySystem {
   private:
