+++ src/core/ecs/system/systems/texture_cube_rendering_entity_system.h
@@ -1,10 +1,10 @@
 #ifndef TEXTURE_CUBE_RENDERING_ENTITY_SYSTEM_H
 #define TEXTURE_CUBE_RENDERING_ENTITY_SYSTEM_H
 
-#include "entity_system.h"
-#include "../component/components/transform3D_component.h"
-#include "../component/components/material_component.h"
-#include "../../math/space_handler.h"
+#include "../entity_system.h"
+#include "../../component/components/transform3D_component.h"
+#include "../../component/components/material_component.h"
+#include "../../../math/space_handler.h"
 
 class TextureCubeRenderingEntitySystem : public EntitySystem {
   private:
